# Optimize connected graph split logic