fixes #3050